remove package apiDefinition
