--- packages/apiDefinition/src/index.ts
@@ -1,26 +1,0 @@
-// Export input schemas
-export {
-	submitContentInput,
-	execSubmissionInput,
-	cancelSubmissionInput,
-} from "./router";
-
-// Export response schemas
-export {
-	submitContentResponse,
-	execSubmissionResponse,
-	cancelSubmissionResponse,
-	getMeResponse,
-} from "./router";
-
-// Export client types (type-only, no runtime code)
-export type {
-	AppRouter,
-	SubmitContentInput,
-	SubmitContentOutput,
-	ExecSubmissionInput,
-	ExecSubmissionOutput,
-	CancelSubmissionInput,
-	CancelSubmissionOutput,
-	GetMeOutput,
-} from "./types";
